Structured frontend and added login
