+++ frontend/src/main.jsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+// import './index.css'
 import ECDHExample from './components/ECDHExample.jsx'
-import SecureChatApp from './components/SecureChatApp.jsx'
+import SecureChatApp from './SecureChatApp.jsx'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
